fix(compression): separate Codex transport and summary progress
